requirements.txt
fastapi[all]
uvicorn
requests
sqlalchemy
psycopg2-binary
python-dotenv
openai
PyMuPDF
python-docx